fix : 1/26, 백준 알고리즘 : [1769번 3의배수 문제]
